requirements.txt
python-telegram-bot==20.*
python-dotenv==1.0.*
notion-client==2.*
google-api-python-client==2.*
google-auth-httplib2==0.*
google-auth-oauthlib==1.*
# openai==1.*  # Temporarily removed due to build issues
